audio: fixed isssues that caused by multich hi pcm output[1/1]
PD# 148436(140637)

hi pcm use i2s device  but outstream is directstream;
mixerstream also use i2s device,so we need control
i2s device to be proper used and so there is no confict.
we degined that if we are at hi pcm mode ,mixerstream will
standby .if mixerstream write need write data , we go to exit
and sleep for proper time.

Change-Id: Iaac5b01e9cfac194a194bb2c3fcea6c76283ef64